+++ yii2-adapter/constants/Field/Elements/Concerns/LegacyConstants.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+namespace CraftCms\Cms\Field\Elements\Concerns;
+
+use craft\base\ElementEventConstants;
+use Deprecated;
+
+/**
+ * @internal
+ * @deprecated 6.0.0
+ */
+trait LegacyConstants
+{
+    use ElementEventConstants;
+}
